# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Agent LLM Research Automation Platform API
  description: |
    Comprehensive API documentation for the Multi-Agent Research Platform.
    
    ## Architecture
    
    The platform consists of:
    - **AI Engine** (Python/FastAPI): Core agent orchestration and LLM processing
    - **Backend** (Node.js/Express): User management, research sessions, and data persistence
    - **Frontend** (React/TypeScript): User interface and visualization
    
    ## Authentication
    
    Most endpoints require JWT authentication via the `x-auth-token` header.
    
    ## Features
    
    - Multi-agent research workflows
    - Real-time streaming chat
    - Memory management
    - Web search integration
    - Export functionality (Markdown, PDF, LaTeX)
    - Token usage tracking
  version: "2.0.0"
  contact:
    name: Research Team
    email: research@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: AI Engine (Python/FastAPI)
  - url: http://localhost:5000
    description: Backend (Node.js/Express)
paths:
...
  /search:
    post:
      tags: ["AI Engine"]
      summary: Unified web search
      description: Search across multiple providers with normalized results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            example:
              query: "autonomous vehicles sensor fusion 2024"
              providers: ["arxiv", "duckduckgo"]
              max_results: 5
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                query: "autonomous vehicles sensor fusion 2024"
                total_results: 42
                providers_used: ["arxiv", "duckduckgo"]
                results: []
        '400':
          description: Invalid providers specified
        '500':
          description: Search failed
...
  /auth/login:
    post:
      tags: ["Authentication"]
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["email", "password"]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: demo@example.com
              password: demo1234
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                user:
                  id: 1
                  email: demo@example.com
                  name: Demo User
        '400':
          description: Invalid credentials
...
  /research:
    post:
      tags: ["Research"]
      summary: Create research job
      description: Submit a new research task
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["task"]
              properties:
                task:
                  type: string
                  description: Research question or task
                paper_url:
                  type: string
                  format: uri
                  description: Optional paper URL to analyze
                depth:
                  type: string
                  enum: ["quick", "deep"]
                  default: "deep"
      responses:
        '201':
          description: Research job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchJob'
    
    get:
      tags: ["Research"]
      summary: List research jobs
      description: Get user's research jobs with pagination
      security:
        - JWTAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Research jobs list
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResearchJob'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /research/{id}:
    get:
      tags: ["Research"]
      summary: Get research job details
      description: Retrieve specific research job with results
      security:
        - JWTAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Research job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchJob'
        '404':
          description: Research job not found
...
  /research/search:
    post:
      tags: ["Research"]
      summary: Search web (proxy)
      description: Proxy to AI Engine search with authentication
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  
  # Chat
...
  /memories:
    get:
      tags: ["Memory"]
      summary: List user memories
      description: Get paginated list of user memories
      security:
        - JWTAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: source
          in: query
          schema:
            type: string
          description: Filter by memory source
      responses:
        '200':
          description: User memories
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    
    post:
      tags: ["Memory"]
      summary: Create memory
      description: Add new memory entry
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["content"]
              properties:
                content:
                  type: string
                  description: Memory content
                source:
                  type: string
                  description: Source of the memory
                metadata:
                  type: object
                  description: Additional metadata
      responses:
        '201':
          description: Memory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
...
  /events:
    post:
      tags: ["Events"]
      summary: Submit execution event
      description: Submit real-time execution event (used by AI Engine)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["research_id", "event_id", "stage", "message"]
              properties:
                research_id:
                  type: integer
                event_id:
                  type: string
                stage:
                  type: string
                severity:
                  type: string
                  enum: ["info", "warn", "error", "success"]
                category:
                  type: string
                  enum: ["stage", "source", "agent", "error"]
                message:
                  type: string
                details:
                  type: object
      responses:
        '200':
          description: Event recorded
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in x-auth-token header
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
    JobStatus:
      type: string
      enum: ["pending", "running", "completed", "failed"]
    ResearchJob:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        task:
          type: string
        paper_url:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        ai_engine_job_id:
          type: integer
          nullable: true
        result_json:
          type: object
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SearchRequest:
      type: object
      required: ["query"]
      properties:
        query:
          type: string
          description: Search query
        providers:
          type: array
          items:
            type: string
          default: ["duckduckgo", "arxiv"]
          description: Search providers to use
        max_results:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: Maximum results per provider
    SearchResult:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        source:
          type: string
        favicon:
          type: string
          format: uri
        thumbnail:
          type: string
          format: uri
        published:
          type: string
    SearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total_results:
          type: integer
        providers_used:
          type: array
          items:
            type: string
        execution_time:
          type: number
    Memory:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        content:
          type: string
        source:
          type: string
          nullable: true
        source_id:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...